feat(game): legacy commands for stats (#236)
* feat(game): implement /set command

* feat(game): implement /set_maxhp command
feat(game): implement /set_maxsp command

* refactor(game): rename SendAsync -> Send

* feat(game): implement /setskill command
feat(game): implement /setskillother command

* feat(game): implement /con+
feat(game): implement /int+
feat(game): implement /str+
feat(game): implement /dex+

* feat(game): implement /mspd command

* feat(game): implement /stat_reset command

* feat(game): implement /all_skills_master command

* refactor(game): implement EPlayerClassGendered

* fix(game): equip shields

* feat(game): implement /full_set command

* feat(game): Implement /m + /mob command

* feat(game): Implement /mob_ld command

* fix(game): /set [Name] exp [Value] causing a stack overflow

* feat(game): Implement /mm command

* feat(game): Implement /group command

* feat(game): Implement /grrandom command

* feat(game): Implement /state command

* feat(game): Implement /gstate command

* feat(game): Implement /purge command

* refactor(game): improved command error messages

* fix: build

+++ src/Tests/Core.Tests/OutgoingPacketTests.cs
@@ -372,7 +372,7 @@ public class OutgoingPacketTests
                 .Concat(obj.CharacterList.SelectMany(c => Array.Empty<byte>()
                     .Concat(BitConverter.GetBytes(c.Id))
                     .Concat(Encoding.ASCII.GetBytes(c.Name))
-                    .Append(c.Class)
+                    .Append((byte)c.Class)
                     .Append(c.Level)
                     .Concat(BitConverter.GetBytes(c.Playtime))
                     .Append(c.St)
